// src/types/course.ts
export interface CourseData {
  key: string;
  thumbnail: string;
  name: string;
  subject: string;
  teacher: string;
  studentsEnrolled: number;
  resources: number;
  mockTests: number;
  status: "Active" | "Inactive";
}
export interface CourseData {
  key: string;
  thumbnail: string;
  name: string;
  subject: string;
  teacher: string;
  studentsEnrolled: number;
  resources: number;
  mockTests: number;
  status: "Active" | "Inactive";
}